Major code reorg with multiple projects
