--- Messages.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace MontyHallSimulator;
-
-/// <summary>
-/// Class that contains messages to be output to the user.
-/// </summary>
-public class Messages
-{
-
-}
