--- MontyHallSimulator.csproj
@@ -1,28 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<OutputType>Exe</OutputType>
-		<TargetFrameworks>net9.0</TargetFrameworks>
-		<GenerateAssemblyInfo>True</GenerateAssemblyInfo>
-		<GenerateDocumentationFile>True</GenerateDocumentationFile>
-		<RepositoryType>git</RepositoryType>
-		<RepositoryUrl>https://github.com/Balein-Fercoadee/MontyHallSimulator</RepositoryUrl>
-		<AssemblyVersion>1.0.2</AssemblyVersion>
-		<FileVersion>1.0.2</FileVersion>
-		<ProductVersion>1.0.2</ProductVersion>
-		<Version>1.0.2</Version>
-	</PropertyGroup>
-
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-		<Optimize>False</Optimize>
-	</PropertyGroup>
-
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-		<Optimize>True</Optimize>
-	</PropertyGroup>
-
-	<ItemGroup>
-	  <PackageReference Include="System.CommandLine" Version="2.0.0-beta6.25358.103" />
-	</ItemGroup>
-
-</Project>
