+++ MontyHallSimulator.sln
@@ -1,14 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.2.32630.192
+VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MontyHallSimulator", "MontyHallSimulator.csproj", "{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimulatorConsole", "SimulatorConsole\SimulatorConsole.csproj", "{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B1C9409E-EE45-4570-8D96-36582E1DF901}"
-	ProjectSection(SolutionItems) = preProject
-		readme.md = readme.md
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimulatorLogic", "SimulatorLogic\SimulatorLogic.csproj", "{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,23 +17,32 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Debug|x64.Build.0 = Debug|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Debug|x86.Build.0 = Debug|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Release|x64.ActiveCfg = Release|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Release|x64.Build.0 = Release|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Release|x86.ActiveCfg = Release|Any CPU
-		{ABF6A9D6-170D-4D26-8E0B-C2913DAEBD7E}.Release|x86.Build.0 = Release|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Debug|x64.Build.0 = Debug|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Debug|x86.Build.0 = Debug|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Release|x64.ActiveCfg = Release|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Release|x64.Build.0 = Release|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Release|x86.ActiveCfg = Release|Any CPU
+		{0CF9EDA6-91D7-4928-8B8D-317CE49CF67B}.Release|x86.Build.0 = Release|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Debug|x64.Build.0 = Debug|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Debug|x86.Build.0 = Debug|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Release|x64.ActiveCfg = Release|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Release|x64.Build.0 = Release|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Release|x86.ActiveCfg = Release|Any CPU
+		{2FD5EC7D-E1D6-4A9B-BAE1-D3B5E93169E0}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {78ADACEF-05F9-4C8C-8B65-9561AD1A8479}
 	EndGlobalSection
 EndGlobal
