+++ SimulatorLogic/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace SimulatorLogic;
+
+public class Class1
+{
+
+}
